+++ src/domain/app/App.tsx
@@ -3,10 +3,6 @@ import { ApolloProvider } from '@apollo/client';
 import { Provider as ReduxProvider } from 'react-redux';
 import { OidcProvider, loadUser } from 'redux-oidc';
 import { MatomoProvider, createInstance } from '@datapunt/matomo-tracker-react';
-import countries from 'i18n-iso-countries';
-import fi from 'i18n-iso-countries/langs/fi.json';
-import en from 'i18n-iso-countries/langs/en.json';
-import sv from 'i18n-iso-countries/langs/sv.json';
 import { HelmetProvider } from 'react-helmet-async';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
@@ -24,10 +20,6 @@ import AppMeta from './AppMeta';
 import AppRoutes from './AppRoutes';
 import AppTitleAnnouncer from './AppTitleAnnouncer';
 import styles from './app.module.css';
-
-countries.registerLocale(fi);
-countries.registerLocale(en);
-countries.registerLocale(sv);
 
 if (process.env.NODE_ENV !== 'production') {
   enableOidcLogging();