+++ src/domain/youthProfile/form/YouthProfileApproverFields.tsx
@@ -1,12 +1,27 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
+import Config from '../../../config';
 import styles from './YouthProfileApproverFields.module.css';
 import ArrayFieldTemplate from '../../../common/components/arrayFieldTemplate/ArrayFieldTemplate';
 import Stack from '../../../common/components/stack/Stack';
 import YouthProfileFormGrid from './YouthProfileFormGrid';
 import FormGroupDescription from './FormGroupDescription';
 import TextInput from './FormikTextInput';
+import Select from './FormikSelect';
+
+const languages = Config.showAdditionalContactLanguagesFeatureFlag
+  ? [
+      'FINNISH',
+      'SWEDISH',
+      'ENGLISH',
+      'FRENCH',
+      'RUSSIAN',
+      'SOMALI',
+      'ARABIC',
+      'ESTONIAN',
+    ]
+  : ['FINNISH', 'SWEDISH', 'ENGLISH'];
 
 type Props = {
   isApproverFieldsRequired?: boolean;
@@ -70,6 +85,15 @@ function YouthProfileApproverFields({
           type="tel"
           labelText={labelRequired('registration.phoneNumber')}
           autoComplete={getAutoComplete('tel')}
+        />
+        <Select
+          className={styles.formInput}
+          name="languageAtHome"
+          label={labelRequired('registration.languageAtHome')}
+          options={languages.map(language => ({
+            label: t(`LANGUAGE_OPTIONS.${language}`),
+            value: language,
+          }))}
         />
       </YouthProfileFormGrid>
       <FormGroupDescription